+++ src/entities/living/mobs/Zombie.java
@@ -130,7 +130,7 @@ public class Zombie extends AbstractMob{
 
 		glPushMatrix();
 
-		glTranslatef((float)x + rotXOffset, (float)y + rotYOffset, 0f);
+		glTranslatef(x + rotXOffset, y + rotYOffset, 0f);
 		glRotatef(rotation, 0.0f, 0.0f, 1.0f);
