+++ src/util/Color4F.java
@@ -23,10 +23,10 @@ public class Color4F {
 	}
 	
 	public Color4F(int red, int green, int blue, int alpha){
-		this.red = 		(float)red / 255f;
-		this.green = 	(float)green / 255f;
-		this.blue = 	(float)blue / 255f;
-		this.alpha = 	(float)alpha / 255f;
+		this.red = 		red / 255f;
+		this.green = 	green / 255f;
+		this.blue = 	blue / 255f;
+		this.alpha = 	alpha / 255f;
 	}
